Add Wireguard server
WireGuard is an extremely simple yet fast and modern VPN that utilizes
state-of-the-art cryptography. It aims to be faster, simpler, leaner,
and more useful than IPsec, while avoiding the massive headache. It
intends to be considerably more performant than OpenVPN.

Since I already have a virtual machine, it just makes sense to use it as
a VPN if the need arise.