Run Vaultwarden natively
Containers are pain in the ass if all you're looking for is to self host
a bunch of tiny services. The reality, however, bites since there lots
of software without proper packages for Linux distributions.

This patch moves Vaultwarden from running in Podman to be run natively
via old plain systemd unit. For that to happen we first need to unpack
Vaultwarden from the docker image to disk, and then run it in isolated
environment.

Part of the reason why I want this is to play a bit with systemd
sandboxing functionality to understand better its pros and cons.