Add Kopia to backup /mnt/data content
Kopia is a fast and secure open-source backup/restore tool that allows
you to create encrypted snapshots of your data and save the snapshots to
remote or cloud storage of your choice. This patch delivers a basic
Ansible role to install and configure Kopia for periodic backups via
systemd timers.

The kopia binary is sandboxed and has only access to bare minimum.