Ensure permissions of the mount point
Turns out that setting DirectoryMode=0700 in the systemd mount unit is
not enough, because that access mode is only used to create a mount
point directory. Once mounted, the mount point acquires access mode of
the mounted resource. This patch makes sure that the permissions are
enforced when the unit is mounted and are back propagated to the mounted
resource (be it a disk or another directory).

This is an important fix since it implies security aspect: the /mnt/data
directory must have 0700 to act as a shield for its data, to make sure
no one else has access to its content.